Corrected calls to DictionaryDataReaderGateway by overloading the constructor to take in a string value with a file location, AND a second constructor with a default value so it can be used with or without a paramater. I then changed and simplified the getter function of getDictionaryFile().
Next I made DictionaryDataReaderGateway implement CreateDictionaryDataAccessObject, so now it plays nicely with GameState.java

I created an interface DictionaryFunctions which implements getDictionaryFile, getDictionaryMap and verifyWord which simply returns true if word is in list, false if it is not. String must be in all uppercase, and have no spaces or symbols, it must also not be null.

(I will create a faster version if time)

From these changes i updated gamestate to include these changes in the setDictionary methods.

I also made some changes to PlaceWordRefillInteractorTest as it wasn't properly setting up the @BeforeAll objects for dictionary, it is now. HOWEVER there are other problems with the test that have nothing to do with the dictionary, as there are no Asserts in the test functions.

Next i updated the Gamestate.java