Still works with Main:) Completed interface functions implementation for DictionaryFunctions in dictionary.
Small changes to strToSet, now strip the String and make it uppercase to prevent issues.
Also moved getCharacterSetDictionary method to DictionaryFunctions

I created DictionaryFunctionsTests, but have them commented out for the time being as they get ignored by the Junit. I was just experimenting to see if an interface could run tests since it was offered by intellij. So I included the file but it serves no immediate purpose as the tests reside in dictionaryTest

Made essential play nice changes to DictionaryDataReaderGateway, DictionaryDataReaderGatewayTest, DictionaryTest, and Main by simply initializing the variable dictionaryGateway with the Arraylist created by the reader, so now it is passed into the game launcher.

Everything passes tests! And it launches without errors:)